+++ platformio.ini
@@ -74,7 +74,7 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
   -<src/lcd/menu/menu_touch_screen.cpp>
   -<src/lcd/menu/menu_tramming.cpp>
   -<src/lcd/menu/menu_ubl.cpp>
-  -<src/lcd/menu/menu_x_axis_twist_compensation.cpp>
+  -<src/lcd/menu/menu_x_twist.cpp>
   -<src/lcd/extui/anycubic_chiron>
   -<src/lcd/extui/anycubic_i3mega>
   -<src/lcd/extui/dgus> -<src/lcd/extui/dgus/fysetc> -<src/lcd/extui/dgus/hiprecy> -<src/lcd/extui/dgus/mks> -<src/lcd/extui/dgus/origin>
@@ -97,7 +97,7 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
   -<src/feature/backlash.cpp>
   -<src/feature/baricuda.cpp> -<src/gcode/feature/baricuda>
   -<src/feature/bedlevel/abl> -<src/gcode/bedlevel/abl>
-  -<src/feature/bedlevel/abl/x_axis_twist_compensation.cpp>
+  -<src/feature/bedlevel/abl/x_twist.cpp>
   -<src/feature/bedlevel/mbl> -<src/gcode/bedlevel/mbl>
   -<src/feature/bedlevel/ubl> -<src/gcode/bedlevel/ubl>
   -<src/feature/bedlevel/hilbert_curve.cpp>
